1 - Nerfed all the post-Provi pre-Polter weapons I could find by 35%, please fix any that I missed. 2 - Aquatic Scourge telegraphs the spin by turning brownish over time before doing it. 3 - Reduced Sentinel, post-Provi Bloodstone bosses and Polter HP by 35%. 4 - Reduced post-ML Aureus HP by 25% because I forgot last time. 5 - Draconic Swarmers spawn more often.

+++ CalPlayer/CalamityPlayer.cs
@@ -5112,7 +5112,7 @@ namespace CalamityMod.CalPlayer
                 }
                 if (silvaMelee && proj.melee)
                 {
-                    double multiplier = (double)player.statLife / (double)player.statLifeMax2;
+                    double multiplier = player.statLife / (double)player.statLifeMax2;
                     damageMult += multiplier * 0.2;
                 }
             }
